MDEV-33408 Initial support for vector DELETE and UPDATE
When the source row is deleted, mark the corresponding node in HNSW
index by setting `tref` to null. An index is added for the `tref` in
secondary table for faster searching of the to-be-marked nodes.

The nodes marked as deleted will still be used for search, but will not
be included in the final query results.

As skipping deleted nodes and not adding deleted nodes for new-inserted
nodes' neighbor list could impact the performance, we now only skip
these nodes in search results.

- for some reason the bitmap is not set for hlindex during the delete so
  I had to temporarily comment out one line

All new code of the whole pull request, including one or several files
that are either new files or modified ones, are contributed under the
BSD-new license. I am contributing on behalf of my employer Amazon Web
Services, Inc.

+++ sql/handler.cc
@@ -5475,6 +5475,11 @@ handler::ha_delete_all_rows()
               m_lock_type == F_WRLCK);
   mark_trx_read_write();
 
+  int err= 0;
+  if ((err= table->open_hlindexes_for_write()) ||
+      (err= table->hlindexes_on_delete_all()))
+    return err;
+
   return delete_all_rows();
 }
 
@@ -8138,7 +8143,7 @@ int handler::ha_write_row(const uchar *buf)
                       { error= write_row(buf); })
 
   MYSQL_INSERT_ROW_DONE(error);
-  if (!error && !((error= table->update_hlindexes())))
+  if (!error && !((error= table->hlindexes_on_insert())))
   {
     rows_stats.inserted++;
     Log_func *log_func= Write_rows_log_event::binlog_row_logging_function;
@@ -8171,6 +8176,9 @@ int handler::ha_update_row(const uchar *old_data, const uchar *new_data)
   DBUG_ASSERT(new_data == table->record[0]);
   DBUG_ASSERT(old_data == table->record[1]);
 
+  if (table->open_hlindexes_for_write())
+    return 1;
+
   uint saved_status= table->status;
   error= ha_check_overlaps(old_data, new_data);
 
@@ -8189,7 +8197,7 @@ int handler::ha_update_row(const uchar *old_data, const uchar *new_data)
                       { error= update_row(old_data, new_data);})
 
   MYSQL_UPDATE_ROW_DONE(error);
-  if (likely(!error))
+  if (likely(!error) && !(error= table->hlindexes_on_update()))
   {
     rows_stats.updated++;
     Log_func *log_func= Update_rows_log_event::binlog_row_logging_function;
@@ -8262,10 +8270,13 @@ int handler::ha_delete_row(const uchar *buf)
   mark_trx_read_write();
   increment_statistics(&SSV::ha_delete_count);
 
+  if (table->open_hlindexes_for_write())
+    return 1;
+
   TABLE_IO_WAIT(tracker, PSI_TABLE_DELETE_ROW, active_index, error,
     { error= delete_row(buf);})
   MYSQL_DELETE_ROW_DONE(error);
-  if (likely(!error))
+  if (likely(!error) && !(error= table->hlindexes_on_delete()))
   {
     rows_stats.deleted++;
     Log_func *log_func= Delete_rows_log_event::binlog_row_logging_function;
